Move: styles from App component to individual screens. Rename: stackNavigator directory to navigators. Add: types to useNavigation hook
