+++ App.tsx
@@ -1,23 +1,14 @@
 import React from 'react';
-import {View, StyleSheet} from 'react-native';
 import {NavigationContainer} from '@react-navigation/native';
 
-import StackNavigator from './src/stackNavigator/StackNavigator';
+import StackNavigator from './src/navigators/StackNavigator';
 
 function App(): JSX.Element {
   return (
     <NavigationContainer>
-      <View style={styles.container}>
-        <StackNavigator />
-      </View>
+      <StackNavigator />
     </NavigationContainer>
   );
 }
 
 export default App;
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-  },
-});
