+++ src/screens/Home.tsx
@@ -1,15 +1,21 @@
 import React from 'react';
-import {Text} from 'react-native';
+import {Text, StyleSheet} from 'react-native';
 import {SafeAreaView} from 'react-native-safe-area-context';
 
-import {ScreenProps} from '../stackNavigator/StackNavigator';
+import {ScreenProps} from '../navigators/StackNavigator';
 
 const Home: React.FC<ScreenProps<'Home'>> = () => {
   return (
-    <SafeAreaView>
+    <SafeAreaView style={styles.container}>
       <Text>Home</Text>
     </SafeAreaView>
   );
 };
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+  },
+});
 
 export default Home;
